+++ components/ContactForm.vue
@@ -1,79 +1,32 @@
-<script setup>
-const props = defineProps({ textColor: String, button: Object });
-
-const showMessage = ref(false);
-
-const submit = () => {
-  showMessage.value = true;
-  setTimeout(() => {
-    showMessage.value = false;
-  }, 4000);
-};
-
-const borderColor = computed(() => {
-  return props.textColor === 'text-white' ? 'border-light' : 'border-dark';
-});
-
-const placeHolderColor = computed(() => {
-  return props.textColor === 'text-white'
-    ? 'placeholder-white'
-    : 'placeholder-dark';
-});
-</script>
-
 <template>
-  <form action="" class="relative" :class="textColor">
+  <form action="">
     <div class="flex flex-col space-y-6">
-      <div class="flex items-center justify-center space-x-1 lg:justify-start">
-        <h4 class="font-display text-xl">Leave us a message</h4>
-      </div>
+      <label for="email" class="sr-only">Your name</label>
       <input
+        id="name"
+        name="name"
         type="text"
         placeholder="Your name"
-        class="rounded-lg border bg-transparent px-6 py-4 transition-all focus:outline-none"
-        :class="[textColor, borderColor, placeHolderColor]"
+        class="rounded-lg bg-white px-6 py-4 text-dark transition-all focus:outline-none"
         required
       />
+      <label for="email" class="sr-only">Your email</label>
       <input
+        id="email"
+        name="email"
         type="email"
         placeholder="Your email"
-        class="rounded-lg border bg-transparent px-6 py-4 transition-all focus:outline-none"
-        :class="[textColor, borderColor, placeHolderColor]"
+        class="rounded-lg bg-white px-6 py-4 text-dark transition-all focus:outline-none"
         required
       />
+      <label for="message" class="sr-only">Your message</label>
       <textarea
+        id="message"
         name="message"
         placeholder="Your message"
-        class="resize-none rounded-lg border bg-transparent px-6 py-4 transition-all focus:outline-none"
-        :class="[textColor, borderColor, placeHolderColor]"
+        class="resize-none rounded-lg bg-white px-6 py-4 text-dark transition-all focus:outline-none"
         required
       ></textarea>
     </div>
-
-    <div v-if="button" class="mt-8 flex justify-center lg:justify-start">
-      <Button :button="button" @click.prevent="submit" />
-    </div>
-
-    <Transition name="fade">
-      <div
-        v-if="showMessage"
-        class="botttom-0 absolute left-0 translate-y-full text-lg"
-        :class="textColor"
-      >
-        Thank you!
-      </div>
-    </Transition>
   </form>
 </template>
-
-<style scoped>
-.fade-enter-active,
-.fade-leave-active {
-  transition: opacity 0.5s ease;
-}
-
-.fade-enter-from,
-.fade-leave-to {
-  opacity: 0;
-}
-</style>