+++ changelog.md
@@ -28,3 +28,4 @@
 - removed `two-col-image-text-section` favor of `two-columns-section`
 - new `faq-section`
 - new `testimonials-section`
+- removed `text_color`, `background_color` and `lead` from `form-section`